backend/requirements.txt
Django==4.0.5
djangorestframework==3.13.1
requests==2.28.1
python-dotenv==0.15.0
psycopg2-binary
django-storages
boto3
gunicorn==20.0.4
djongo==1.3.6
pymongo[srv]==3.12.3
djangorestframework-simplejwt
uuid
django-cors-headers
graphene-django
graphene==3.2.2
graphene_django==3.0.0
django-filter
